Fix wearos font issue
