+++ components/wearable/setup/impl/src/main/java/com/flipperdevices/wearable/setup/impl/composable/ComposableFindPhone.kt
@@ -1,7 +1,5 @@
 package com.flipperdevices.wearable.setup.impl.composable
 
-import androidx.compose.foundation.focusable
-import androidx.compose.foundation.gestures.scrollBy
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
@@ -11,16 +9,10 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.input.rotary.onRotaryScrollEvent
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
@@ -32,11 +24,11 @@ import com.flipperdevices.core.ui.ktx.clickableRipple
 import com.flipperdevices.core.ui.theme.FlipperThemeInternal
 import com.flipperdevices.core.ui.theme.LocalTypography
 import com.flipperdevices.wearable.core.ui.components.ComposableFlipperButton
+import com.flipperdevices.wearable.core.ui.components.ComposableWearOsScrollableColumn
 import com.flipperdevices.wearable.setup.impl.R
 import com.flipperdevices.wearable.setup.impl.model.FindPhoneState
 import com.flipperdevices.wearable.setup.impl.viewmodel.FindPhoneViewModel
 import com.google.android.horologist.compose.layout.fillMaxRectangle
-import kotlinx.coroutines.launch
 
 @Composable
 fun ComposableFindPhone(
@@ -44,25 +36,7 @@ fun ComposableFindPhone(
     modifier: Modifier = Modifier,
     findPhoneViewModel: FindPhoneViewModel = viewModel()
 ) {
-    val columnScrollState = rememberScrollState()
-    val coroutineScope = rememberCoroutineScope()
-    val focusRequester = remember { FocusRequester() }
-
-    Column(
-        modifier = modifier
-            .verticalScroll(columnScrollState)
-            .fillMaxRectangle()
-            .onRotaryScrollEvent {
-                coroutineScope.launch {
-                    columnScrollState.scrollBy(it.verticalScrollPixels)
-                }
-                true
-            }
-            .focusRequester(focusRequester)
-            .focusable(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
+    ComposableWearOsScrollableColumn(modifier) {
         val findPhoneModel by findPhoneViewModel.getFindPhoneModelFlow().collectAsState()
         val localFindPhoneModel = findPhoneModel
 
@@ -74,9 +48,6 @@ fun ComposableFindPhone(
                 findPhoneViewModel::checkPhone
             )
         }
-    }
-    LaunchedEffect(Unit) {
-        focusRequester.requestFocus()
     }
 }
