+++ components/wearable/sync/wear/impl/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
     implementation(projects.components.core.ui.theme)
     implementation(projects.components.core.ui.navigation)
     implementation(projects.components.wearable.core.ui.ktx)
+    implementation(projects.components.wearable.core.ui.components)
 
     implementation(projects.components.bridge.dao.api)
     implementation(projects.components.wearable.emulate.wear.api)
